Test multiple resolvers via matrix
Runs the `test` job against multiple Stack resolvers using a matrix strategy.

Motivated by #6, to exercise each case of the preprocessor directive that tests the version of `persistent`. LTS 18.0 pulls in the earlier version while the current 18.25 tests the later version.

Also, pins the version of `stack-action` and bumps it and `stack-cache-action` to latest version.