Set `STACK_YAML` env for release
`stack-upload-action` was updated recently to add this env variable. The main `stack.yaml` should be used for release purposes. Since `working-directory` is in each package of the monorepo, `stack` needs to look up a directory for this file.

Using a relative path is a bit fragile but there doesn't appear to be a variable used to refer to the project root to avoid it.